Makes v2 map, see description
One of my playtesters complained about how the elevator worked, especailly when going down the hole, you'd land on the lift which would send you right back up. I experimented a bit with lowering the lift speed, shrinking the size of the lift and expanding the hole in which it would go through. Lowering the speed of the lift I found was not fun. While it did somewhat mitigate the issue from falling through the hole, it would clearly be a problem going back up. IF someone was trying to escape by going up the lift, their pursuer would simply shoot them off. Even if they were in a chase, it still would be an extremely slow process that would prevent fast-paced pvp to occur. It did not flow well at all. While shrinking the lift and opening the hole up did mitigate a bit, I believe players would not expect having to jump off the lift. IF they were to walk off the lift after riding up it, they would just fall back down again, which seems incredibly frustrating. At that point, I think players would have just given up. The gap is still there, but not big enough to fall through when the lift is at the top. When I was looking at the options of the object, I noticed a bool called "lowTriggerVolume". I had no idea what it did, so I Googled it, and Google had no idea what it did, so I checked it and compiled and booted up the map. It actually does exactly what I want it to do. When the player falls on top of it, it doesn't activate for a few seconds. If they do jump on top of it from the bottom or sides, it activates faster. Serendipitous! V3 will have changes that reflect feedback.