Updates level, adds pickups, see desc
I decided to keep a tally of how many pickups I have in the level. I've scaled back quite a bit so hopefulyl there isn't too little now, but we'll see after testing.I put pickups, such as health and ammo, outside the main focal point but on paths leading out, so if players need to recoup their losses, they can see the pickups. Taking a lap around the level will also lead to one of two gun pickups. The idea is that players will fight in the main focal point, then use paths to escape or chase other players, collect ammo and health.  The only weapons I've added in the super shotgun and the base nail gun, while keeping the rocket launcher in the middle. The launcher seemed to be a popular move, so I'll keep that. The nailgun is part of the path that connects the lower middle to the top middle. The super shotgun is in the point beterrn the teleporter and the dual stairway. There are 4 healthkits. Two upstairs and two downstairs. There are four spike pickups, two on both floors. There are three shotgun pickups. 1 upstairs, and 2 downstairs. I think this level will take on more of a prison vibe with cages all around.